Rename, restructure, add geofence on socket event
