+++ client/src/app.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter, Route } from 'react-router-dom';
 import store from './store';
-import tracker from './tracker';
+import listeners from './listeners';
 import App from './components/App';
 
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
