+++ client/src/components/NewLote.js
@@ -9,7 +9,7 @@ import MenuItem from 'material-ui/MenuItem';
 import MapContainer from './MapContainer';
 import { Card, CardMedia } from 'material-ui/Card';
 import Checkbox from 'material-ui/Checkbox';
-import socket from '../socket'; 
+import socket from '../socket';
 
 class NewLote extends React.Component {
   constructor(props) {
@@ -22,7 +22,6 @@ class NewLote extends React.Component {
 
     this.handleSubmit = this.handleSubmit.bind(this);
     this.placeSubmit = this.placeSubmit.bind(this);
-    this.placeSearch = this.placeSearch.bind(this);
     this.placeRef = this.placeRef.bind(this);
     this.handleLockToggle = this.handleLockToggle.bind(this);
     this.handleRecipientChange = this.handleRecipientChange.bind(this);
@@ -50,7 +49,7 @@ class NewLote extends React.Component {
   }
 
   handleSubmit(event) {
-    event.preventDefault(); 
+    event.preventDefault();
 
     let loteInfo = {
       senderId: this.props.profile.id,
@@ -62,14 +61,14 @@ class NewLote extends React.Component {
       longitude: this.props.lotecation.lng || this.props.userLocation.lng,
       latitude: this.props.lotecation.lat || this.props.userLocation.lat
     };
-    
+
     socket.emit('send message', loteInfo, (err, msg) => {
       if (err) {
         console.log (err);
       } else {
         console.log (msg);
         this.props.setActiveMessage('');
-        // this.props.getLotes(this.props.profile.id), 
+        // this.props.getLotes(this.props.profile.id),
         this.props.history.push('/lotes');
       }
     });
@@ -102,10 +101,6 @@ class NewLote extends React.Component {
     console.log(event);
   }
 
-  placeSearch(event) {
-    console.log('hello');
-  }
-
   render() {
     let p = this.props;
     const mapProps = {
@@ -127,9 +122,9 @@ class NewLote extends React.Component {
             {this.props.contacts.map((contact, i) => {
               return (
                 contact.receiver_id !== this.props.profile.id &&
-                  <MenuItem 
-                  key={ i + 1 } 
-                  value={ contact.receiver } 
+                  <MenuItem
+                  key={ i + 1 }
+                  value={ contact.receiver }
                   primaryText={ contact.receiver.display ? contact.receiver.display : contact.receiver.email }
                   />
               );
@@ -137,40 +132,39 @@ class NewLote extends React.Component {
           </DropDownMenu>
           <form className="lote-form" ref="loteForm" onSubmit={ this.handleSubmit }>
             <TextField hintText="Enter a message" multiLine={true} rows={1} ref="message" type="text" name="message" value={ this.props.activeMessage } onChange={ (event) => this.props.setActiveMessage(event.target.value) } />
-            <form onSubmit={this.placeSubmit}>
-              <TextField id="locationSearch" ref={this.placeRef} />
-              <RaisedButton labelColor='#ffffff' backgroundColor='#48d09b' className="submitButton" label="Search" type="submit" onTouchTap={ this.placeSearch }/>
-            </form>
-            <div className="lote-form-label">
-              <label className="lote-form-label">Radius</label>
-                <DropDownMenu ref="radius" value={this.state.radius} onChange={ this.handleRadiusChange } openImmediately={ false }>
-                  <MenuItem value={20} primaryText='20 meters'/>
-                  <MenuItem value={100} primaryText='100 meters'/>
-                  <MenuItem value={500} primaryText='500 meters'/>
-                  <MenuItem value={2500} primaryText='2,500 meters'/>
-                  <MenuItem value={10000} primaryText='10,000 meters'/>
-                </DropDownMenu>
-              
-            </div>
-            <div>
-              <Checkbox 
-                label='Location-Locked' 
-                style={{width: 'initial', margin: 'auto', paddingRight: 12}} 
-                labelStyle={{width: 'initial'}} 
-                checked={ this.state.lock } 
-                onCheck={ this.handleLockToggle } 
-              />
-            </div>
-            <div>
-              <RaisedButton 
-                labelColor='#ffffff' 
-                backgroundColor='#48d09b' 
-                className="submitButton" 
-                label="Submit" 
-                onTouchTap={ this.handleSubmit }
-              />
-            </div>
           </form>
+          <form onSubmit={this.placeSubmit}>
+            <TextField id="locationSearch" ref={this.placeRef} />
+          </form>
+          <div className="lote-form-label">
+            <label className="lote-form-label">Radius</label>
+              <DropDownMenu ref="radius" value={this.state.radius} onChange={ this.handleRadiusChange } openImmediately={ false }>
+                <MenuItem value={20} primaryText='20 meters'/>
+                <MenuItem value={100} primaryText='100 meters'/>
+                <MenuItem value={500} primaryText='500 meters'/>
+                <MenuItem value={2500} primaryText='2,500 meters'/>
+                <MenuItem value={10000} primaryText='10,000 meters'/>
+              </DropDownMenu>
+
+          </div>
+          <div>
+            <Checkbox
+              label='Location-Locked'
+              style={{width: 'initial', margin: 'auto', paddingRight: 12}}
+              labelStyle={{width: 'initial'}}
+              checked={ this.state.lock }
+              onCheck={ this.handleLockToggle }
+            />
+          </div>
+          <div>
+            <RaisedButton
+              labelColor='#ffffff'
+              backgroundColor='#48d09b'
+              className="submitButton"
+              label="Submit"
+              onTouchTap={ this.handleSubmit }
+            />
+          </div>
         </Card>
       </div>
     );
