--- client/src/lib/pointTree.js
@@ -1,4 +1,0 @@
-import rbush from 'rbush';
-// create and export to hold in global state
-export default rbush();
-
