+++ client/src/reducers/lotes.js
@@ -4,9 +4,9 @@ const lotes = (state = [], action) => {
     console.log('adding lotes to store');
     return action.lotes;
   case 'ONE_LOTE' :
-  	let newState = state.slice(); 
-  	newState.push(action.lote); 
-  	return newState; 
+    let newState = state.slice();
+    newState.push(action.lote);
+    return newState;
   default:
     return state;
   }
